storage: move fe,be to a separate file
