+++ storage/Makefile
@@ -1,6 +1,6 @@
-#Copyright (C) 2022 Intel Corporation
-#SPDX-License-Identifier: Apache-2.0
-#
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (C) 2022 Intel Corporation
+# Copyright (c) 2022 Dell Inc, or its subsidiaries.
 
 shellcheck:
 	../scripts/run-shellcheck.sh
