+++ storage/README.md
@@ -31,7 +31,7 @@ docker run --rm -it -v `pwd`:/app -w /app golang:alpine go get github.com/opipro
 Run example server (not for production) manually
 
 ```bash
-   docker run --rm -it -v `pwd`:/app -w /app -p 50051:50051 golang:alpine go run jsonrpc.go server.go
+   docker run --rm -it -v `pwd`:/app -w /app -p 50051:50051 golang:alpine go run jsonrpc.go frontend.go backend.go middleend.go server.go
 ```
 
 Run example client (not for production) manually
